+++ langs/tree-sitter-langs-build.el
@@ -1,4 +1,4 @@
-;;; tree-sitter-langs.el --- Language definitions for tree-sitter -*- lexical-binding: t; coding: utf-8 -*-
+;;; tree-sitter-langs-build.el --- Building grammar bundle -*- lexical-binding: t; coding: utf-8 -*-
 
 ;; Copyright (C) 2020 Tuấn-Anh Nguyễn
 ;;
@@ -6,10 +6,7 @@
 
 ;;; Commentary:
 
-;; This is a convenient bundle of language definitions for `tree-sitter'. It
-;; serves as an interim distribution mechanism, until `tree-sitter' is
-;; widespread enough for language major modes to include the definitions on
-;; their own.
+;; This file contains functions to obtain and build `tree-sitter' grammars.
 
 ;;; Code:
 
@@ -19,14 +16,30 @@
 (require 'url)
 (require 'tar-mode)
 
-(require 'tree-sitter-cli)
-(require 'tree-sitter-load)
-
 (eval-when-compile
   (require 'subr-x)
-  (require 'pcase))
+  (require 'pcase)
+  (require 'cl-lib))
 
-(defconst tree-sitter-langs--version "0.0.7")
+(defconst tree-sitter-langs--suffixes '(".dylib" ".dll" ".so")
+  "List of suffixes for shared libraries that define tree-sitter languages.")
+
+(defconst tree-sitter-langs--dir
+  (file-name-directory (locate-library "tree-sitter-langs")))
+
+(defconst tree-sitter-langs--bin-dir
+  (file-name-as-directory
+   (concat tree-sitter-langs--dir "bin")))
+
+(defconst tree-sitter-langs--version
+  (let ((main-file (locate-library "tree-sitter-langs.el")))
+    (unless main-file
+      (error "Could not find tree-sitter-langs.el"))
+    (with-temp-buffer
+      (insert-file-contents-literally main-file)
+      (unless (re-search-forward ";; Version: \\(.+\\)")
+        (error "Could not determine tree-sitter-langs version"))
+      (match-string 1))))
 
 (defconst tree-sitter-langs--os
   (pcase system-type
@@ -52,7 +65,7 @@ If VERSION and OS are not spcified, use the defaults of
           (tree-sitter-langs--bundle-file ".gz" version os)))
 
 ;;; A list of (LANG-SYMBOL VERSION-TO-BUILD &optional PATHS REPO-URL).
-(defvar tree-sitter-langs-repos
+(defconst tree-sitter-langs-repos
   '((agda       "v1.2.1")
     (bash       "v0.16.0")
     (c          "v0.16.0")
@@ -78,14 +91,10 @@ If VERSION and OS are not spcified, use the defaults of
     (typescript "v0.16.1" ("typescript" "tsx")))
   "List of language symbols and their corresponding grammar sources.")
 
-(defconst tree-sitter-langs--grammars-dir
-  (let ((this-dir (file-name-directory (locate-library "tree-sitter-langs"))))
-    (file-name-as-directory
-     (concat (if (file-exists-p (concat this-dir "Cargo.toml"))
-                 this-dir
-               (file-name-directory
-                (directory-file-name this-dir)))
-             "grammars")))
+(defconst tree-sitter-langs--repos-dir
+  (file-name-as-directory
+   (concat (file-name-directory (locate-library "tree-sitter-langs"))
+           "repos"))
   "Directory to store grammar repos, for compilation.")
 
 (defun tree-sitter-langs--source (lang-symbol)
@@ -110,7 +119,9 @@ If BUFFER is nil, `princ' is used to forward its stdout+stderr."
                      `(:buffer ,tree-sitter-langs--out)
                    `(:filter (lambda (proc string)
                                (princ string)))))
-         (proc (apply #'make-process (append base output)))
+         (proc (let ((process-environment (cons (format "TREE_SITTER_DIR=%s" tree-sitter-langs--dir)
+                                                process-environment)))
+                 (apply #'make-process (append base output))))
          (exit-code (progn
                       (while (not (memq (process-status proc)
                                         '(exit failed signal)))
@@ -128,7 +139,7 @@ If there's no tag, return \"origin/master\"."
      (pcase-let*
          ((`(,lang-symbol . _) desc)
           (lang-name (symbol-name lang-symbol))
-          (default-directory (concat tree-sitter-langs--grammars-dir
+          (default-directory (concat tree-sitter-langs--repos-dir
                                      (format "tree-sitter-%s" lang-name))))
        `(,lang-symbol ,(or (magit-get-current-tag)
                            "origin/master"))))
@@ -157,7 +168,7 @@ Requires git and tree-sitter CLI."
          (lang-name (symbol-name lang-symbol))
          (dir (if source
                   (file-name-as-directory
-                   (concat tree-sitter-langs--grammars-dir
+                   (concat tree-sitter-langs--repos-dir
                            (format "tree-sitter-%s" lang-name)))
                 (error "Unknown language `%s'" lang-name)))
          (repo (plist-get source :repo))
@@ -201,11 +212,11 @@ The bundle includes all languages declared in `tree-sitter-langs-repos'."
         (let* ((tar-file (concat (file-name-as-directory
                                   (expand-file-name default-directory))
                                  (tree-sitter-langs--bundle-file)))
-               (default-directory (tree-sitter-cli-bin-directory))
+               (default-directory tree-sitter-langs--bin-dir)
                (tree-sitter-langs--out (tree-sitter-langs--buffer "*tree-sitter-langs-create-bundle*"))
                (files (seq-filter (lambda (file)
                                     (when (seq-some (lambda (ext) (string-suffix-p ext file))
-                                                    tree-sitter-load-extensions)
+                                                    tree-sitter-langs--suffixes)
                                       file))
                                   (directory-files default-directory)))
                ;; Disk names in Windows can confuse tar, so we need this option. BSD
@@ -224,13 +235,13 @@ The bundle includes all languages declared in `tree-sitter-langs-repos'."
 ;;;###autoload
 (defun tree-sitter-langs-install (&optional version os keep-bundle)
   "Download and install the specified VERSION of the language grammar bundle.
-If VERSION and OS are not spcified, use the defaults of
+If VERSION and OS are not specified, use the defaults of
 `tree-sitter-langs--version' and `tree-sitter-langs--os'.
 
 The download bundle file is deleted after installation, unless KEEP-BUNDLE is
 non-nil."
   (interactive)
-  (let ((dir (tree-sitter-cli-bin-directory)))
+  (let ((dir tree-sitter-langs--bin-dir))
     (unless (file-directory-p dir)
       (make-directory dir t))
     (let ((default-directory dir)
@@ -279,5 +290,5 @@ non-nil."
       (url-copy-file url gz-file)
       (dired-compress-file gz-file))))
 
-(provide 'tree-sitter-langs)
-;;; tree-sitter-langs.el ends here
+(provide 'tree-sitter-langs-build)
+;;; tree-sitter-langs-build.el ends here